pyproject.toml
[tool.poetry]
name = "Contacts API"
version = "0.2.0"
description = "Based on FastAPI."
authors = ["Oleksandr Horbatiuk <oleksandr.v.horbatiuk@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
